+++ launcher/HowlingWhispersLauncher.cs
@@ -8,8 +8,8 @@ using System.Windows.Forms;
 [assembly: AssemblyDescription("Windows launcher for The Howling Whispers")]
 [assembly: AssemblyCompany("Patina Works")]
 [assembly: AssemblyProduct("The Howling Whispers")]
-[assembly: AssemblyVersion("0.4.2.7")]
-[assembly: AssemblyFileVersion("0.4.2.7")]
+[assembly: AssemblyVersion("0.4.2.8")]
+[assembly: AssemblyFileVersion("0.4.2.8")]
 
 internal static class HowlingWhispersLauncher
 {